Moved agenda and zekering vue components to separate directory
