+++ resources/assets/vue/components/agenda/components/AgendaItem.vue
