+++ resources/assets/vue/components/agenda/constants.js
