+++ resources/assets/vue/components/agenda/event-bus.js
