+++ resources/assets/vue/components/agenda/store/index.js
@@ -1,6 +1,6 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
-import modules from './modules/index';
+import modules from './modules';
 
 Vue.use(Vuex);
