+++ resources/assets/vue/components/agenda/store/modules/agenda.js
