+++ resources/assets/vue/components/agenda/store/modules/index.js
