+++ resources/assets/vue/components/zekeringen/event-bus.js
@@ -1,0 +1,5 @@
+import Vue from 'vue';
+
+const EventBus = new Vue();
+
+export default EventBus;
